graphics/vulkan-extension-layer: add new port
There are some extensions and features in Vulkan that are not
available everywhere due to various reasons. While not available, some
extensions are capable of being done as layer and mapping to any
Vulkan implementation regardless of it supporting the desired
functionality or not.

https://github.com/KhronosGroup/Vulkan-ValidationLayers
(cherry picked from commit 64e16df573cf569f1d1056963fff072e990a2c3b)