devel/llvm12: Add patches to enable kernel sanitizers on FreeBSD
I use CROSS_TOOLCHAIN quite frequently, but can't do so when building
kernels with KASAN or KMSAN enabled since these patches didn't make it
into LLVM 12.  This diff just backports them to the port. I believe this
is low-risk and thus is acceptable for the port, as opposed to waiting
for LLVM 13.

Differential Revision:	https://reviews.freebsd.org/D31316