games/anki: Update to 2.1.44
Since the 2.1.36 release the build system has changed significantly
again and devel/bazel is now used by upstream for the entire build
process.

With this change, a lot of bazel rules and also some translation files
needs to be downloaded from various repositories during the buildtime.

In order to still make Anki compatible with the Ports framework, the
following approach is used:

  1. Prefetch all the required files before buildtime.
  2. Create and setup a few local bazel repositories along with a set
     of custom bazel rules and to provide Node, Protobuf and Yarn from
     the Ports tree.
  3. Compile the Rust library and place the output files to the
     respective local bazel repository for further processing.
  4. Build the rest via devel/bazel

From a technical point of view, Anki should also be able to be built on
other 64-bit architectures (but requires further patching of a few bazel
rules), as soon as devel/bazel can also be built for those platforms.

The option NLS is gone because the translations are now compiled
directly into the Rust library and they only generate a small overhead.

Changelog since 2.1.35 (too many changes/fixes to list them separately):

https://changes.ankiweb.net/

Reviewed by:	tcberner
MFH:		2021Q3
Differential Revision:	https://reviews.freebsd.org/D31358