editors/vim: Explicitly disable libcanberra support
Vim will link against libcanberra if it's installed. It really doesn't
serve any particular purpose for a text editor, so it makes sense for
us to just disable it explicitly.

This change existed before, but was hidden in the GNOME option and got
removed when libgnomeui was purged.

PR:		257565
Reported by:	Philipp Ost