lang/gcc11-devel: Update to the 20210731 snapshot of GCC 11.2.1
This is the first snapshot after the release of GCC 11.2.

It fixes one warning and brings four back ports for the middle end,
one for the RTL optimizers, three for the x86 and powerpc (nee rs6000)
back ends each, four for C++, one for C, and three for Fortran, GOMP
and libstdc++ each.