emulators/wine-devel: Fix build when devel/libsysinfo is present
When devel/libsysinfo is present, <sys/sysinfo.h> exists which
misleads Wine into enabling some code that invokes the sysinfo
function. That one isn't actually available, though, without
additional linking parameters (and we don't want to add another
dependency at this point anyway), so properly test for sysinfo()
before using it.

This is a patch of mine that was accepted upstream and should
come in via Wine 6.15.

PR:		257398