devel/binutils: Add fix for CVE-2021-3487
The CVE is fixed in main in a0e752df8013. Merging that would mean merging other
changes to other ports and doing more exp-runs, so we just backport the fix in
the quarterly branch to avoid too much disruption.

VuXML entry to be handled in PR 256133.

PR:	255368, 251385
Reported by:	diizzy@
Security:	CVE-2021-3487