www/py-puppetboard: Improve sample settings file
Install a new settings.py sample file which contain a better description
of it's expected content and has a file name that better match what it
is used for.

Copying default_settings.py from the project was a leftover from a
previous version of PuppetBoard which does not make much sense nowadays.