Mk/bsd.port.mk: Clarify UNAME_r failure.
An example of this happening is with recent Poudriere changes:
- Poudriere <= 3.3 used to set UNAME_r in /etc/login.conf's environment
  during *jail creation*.
- Poudriere >= 3.4 only sets it in the environment during *jail
  runtime*.

Creating a jail with Poudriere 3.4 and then using Poudriere 3.3 on it
would fail to have UNAME_r properly set.  It would read OSVERSION from
/usr/include/sys/param.h to find the installed files release.  The ports
framework, and builds, need `uname` to be correct for the *jail/chroot*
and not be of the running kernel.

Add a footshoot enabler while here which should not be used for any
reason as it can create packages targetting the release of the running
kernel rather than the chroot/jail's expected version.