devel/llvm-devel: new snapshot post 13 branch
Update to a new snapshot after the branch point for the 13.0.0 release.

This snapshot includes a new scanbuild written in python.  I've not
tested it and it likely requires further refinement.

Allow LOCALBASE/llvm-devel and LOCALBASE/bin to be on different file
systems. [0]

PR:		244439 [0]
Reported by:	eugen [0]
Sponsored by:	DARPA