games/linux-dwarffortress: Update to 0.47.05
The non-Premium (e.g. not the one that will eventually be on Steam)
version of Dwarf Fortress gets occasional releases. The release
notes can be found on
	http://www.bay12games.com/dwarves/
if you scroll down a long way; there is no separate release page.
Some major and minor bug fixes, improved overall Dwarfishness.

This version works with 0.47.04 save files, although you might
not get the bugfixes if continuing in such a world.