games/gnugo: drop bogus personal attribution from the port description
It had been added in r22849 during routinely version update and is the
name of the submitter of PR 13713, which means it should not have been
committed in the first place as we do not append names of contributors
to port descriptions.  Replace it with canonical WWW line which points
to the project's home page.