net/freerdp: add workaround for focus bug of WM
Some WMs (for example fvwm and fluxbox) intermittently steal focus
during attempts to drag something with the mouse on the remote machine,
effectively making it impossible to drag. This behavior seems to also
depend on the configuration of the WM.

Add an optional patch to work around this problem, activated by the new
option BROKENFOCUS. Note that this workaround may lead to a wrong state
of keyboard modifiers after focus changes, so only enable if if you
really need to.

PR:	254908
Approved by:	VVD <vvd@unislabs.com> (maintainer)