devel/RStudio: Fix bug in RStudio-server that gmake isn't in the path
When R modules are installed gmake is needed, but it wasn't in the path
in the server flavor of the port. The patch adds the gmake symlink to
the directory which is in the path.

A better fix will be provided later, pending fix in RStudio, see Makefile.

Also adjust poudriere.conf instructions.

(cherry picked from commit 4a933f3420f6dad303c07a4b90c6d2d591f84b98)