devel/ocaml-ocamlbuild: fix install in presence of `devel/ocaml-findlib'
If ocamlfind(1) is installed, its output of `printconf destdir' is used
as OCAMLBUILD_LIBDIR, which does not match OCAML_LIBDIR and is actually
not a single directory, but colon-separated path list.  Specify correct
destination explicitly to avoid this type of side effects.  While at it,
hook the test suite to our framework.

PR:		238905
Reported by:	John Hein