net/openldap24-server: Make DEBUG default.
In 616f384d5cce7f52a9f5a2164d405f4185bbcab9 the port was converted
to use the option framework's handling of options.  Previously,
when "DEBUG" is unchecked, the port will not pass --disable-debug
and because debug was enabled by default in upstream, the option
is effectively saying build with debug anyways.  After the conversion,
we are now correctly building a non-debug build which will also remove
syslog support.

Restore the historical behavior and build with debug option by default
as it's what expected by upstream and the option never worked on FreeBSD
until the recent conversion actually fixed the option to do what it was
intended to do.

PR:		ports/257884