devel/freebsd-sysroot: Add port
This port is intended as a helper for lang/rust-bootstrap and
provides the necessary sysroots to cross-compile the bootstraps.

The sysroots provide target architecture base headers, base libraries,
and convenient cc and c++ wrapper scripts.  Always for the lowest
supported FreeBSD version.

The sysroot setup from lang/rust-bootstrap can later be swapped out
for this port.  It might also be useful for others things.

Compared to lang/rust-bootstrap all sysroots are moved to at least
12.2-RELEASE since 11.4-RELEASE is EOL after 2021-09-30 and Rust
1.56 is scheduled for 2021-10-21.