emulators/virtualbox-ose-additions: Assign default value to pointer
In the virtualbox virtual filesystem code we ship as a patch some
code in an else block is commented out. This produces a code path
in which a pointer variable is dereferenced in an unassigned state,
causing random crashes.

Lacking a better fix, give a default value of NULL to the pointer,
which at least avoids the random pointer dereference issue.

PR:		255386