sysutils/plasma5-ksystemstats: new port for system resource usage
Plasma Systemstats is a successor to (lib)KSysGuard in the way
that Plasma System Monitor is a successor to KSysGuard. The
background idea is that system statistics collection (resource
usage) should be low-overhead and pluggable and the UI in the
application flexible. KSysGuard achieved that with 2010-era
technologies, but time and technology moves on.

Note, though, that the backend -- the actual system-statistics-
collection -- on FreeBSD is missing collection of network
activity (entirely) and memory use (is wrong). System monitor
is **not yet** a replacement for KSysGuard on FreeBSD, but the
CPU monitoring -- which might be the most important and graphically
appealing bit on a desktop machine -- is much improved.

This commit introduces the new port, adds it to USE_KDE,
and makes it a dependency of system-monitor, to ease the
eventual replacement (once the issues mentioned above are
cleared up) of old by new.