framework: new USES trigger
Triggers are a new feature from pkg 1.17 (stable in 1.17.2)
they are script that are executed at the end of a transaction if
anything triggered them.

To trigger them a files should have been installed in a location which
matches predefined pattern (glob, regexp, or exact matching)

When a package is installed in a cross root installation for example:
pkg -r XXX install something
and if it triggers a trigger, then the execution will be deferred.

The pkg trigger command will execute the deferred triggers

Reviewed by: tcberner
Differential Revision: https://reviews.freebsd.org/D31960