devel/py-aws-requests-auth: Add py-aws-requests-auth 0.4.3
This package allows you to authenticate to AWS with Amazon's signature version 4
signing process with the python requests library.

Conceivably, the authentication class is flexible enough to be used with any AWS
service, but it was initially created to interface with AWS Elasticsearch
instances.

WWW: https://github.com/DavidMuller/aws-requests-auth