devel/cmake: fix CPack .pkg support, jump to 3.21.3
Update the patch for FreeBSD .pkg support in CPack to handle
libpkg 1.17 and later's use of .pkg as package-suffix rather
than .tgz. The PR below refers to this fix. I've kept code
for older libpkg in case there are systems with weirdly
mismatched current CMake and last-quarter's libpkg.

The patch is still being upstreamed. The PR mentioned below
is for this fix.

While here, update to 3.21.3.  I've decided to -- unusual,
for sure -- bump CMake without the usual exp-run, because:

A regression in CMake 3.21.2 prompted a quick release of
3.21.3 from upstream. The regression applies to the AMD ROCm
platform, so that's not us. Other fixes in 3.21.3 apply to supported
MSVC versions -- again, not us.

There is new support for Boost 1.77 -- not in our ports yet.

In other words, the git diff on CMake sources does not show
anything that could affect us (famous last words). Release notes are at
	https://cmake.org/cmake/help/v3.21/release/3.21.html#id3

PR:		257615
Reported by:	Peter Laursen