USES=ncurses: expose variable with the name of the lib to link against
ncurses from ports and futur ncurses in base differs from current
ncurses in base by the fact it is not providing a giant libncurses.so
but 2 libraries: libtinfo and libncurses, those information often needs
to be reported to the consumer port.

while here remove a now useless test for support of widechar ncurses in
base, all supported version of freebsd have it now.