misc/mc: improve the check whether installed diff(1) is GNU or BSD
In 2017, GNU diff had been replaced by BSD licensed diff in our base,
but some features of GNU diff had not been implemented, particularly,
the ``group format'' family of options.  Instead of checking for the
${OSVERSION} value, check if installed diff(1) implementation is GNU
or BSD one based on the --version output.

Because conditions are subject to immediate expansion, we cannot use
${DIFF} and ${GREP} variables; however, literal ``diff'' is actually
correct in this particular case as user's preferred diff(1) does not
necessarily point to ``/usr/bin/diff'' which ${DIFF} expands to.

While here, amend the EDITOR option description so it clearly refers
to the internal *text* editor.