archivers/upx: fix build with clang 13
During an exp-run for llvm 13 (see bug 258209), it turned out that
archivers/upx fails to build with clang 13:

p_wcle.cpp:739:27: error: variable 'n' set but not used [-Werror,-Wunused-but-set-variable]
    unsigned count,object,n,r;
                          ^
1 error generated.

This is because clang 13 now has a -Wunused-but-set-variable warning
similar to gcc's, and it is enabled under -Wall.

The p_wcle.cpp file has two instances where the 'n' variable is used for
debugging purposes, but the first instance is marked with UNUSED(n). The
second is not, triggering this warning. Fix it by also marking the
second instance with UNUSED(n).

PR:		258394
Approved by:	maintainer timeout (3 weeks)