devel/py-pdm: Add py-pdm 1.8.5
PDM is meant to be a next generation Python package management tool. It was
originally built for personal use. If you feel you are going well with Pipenv or
Poetry and don't want to introduce another package manager, just stick to it.
But if you are missing something that is not present in those tools, you can
probably find some goodness in pdm.

WWW: https://github.com/pdm-project/pdm