devel/llvm13: Disable OpenMP amdgcn bit code
Due to a bug in the CMakefiles these build for aarch64 and maybe
ppc64le.  This resulted in plist breakage for aarch64.

It would be straightforward to enable them for amd64, but I have no way
to test them and the current CMake bits build them for linux targets so
I suspect they won't work.

PR:		258706
Reviewed by:	jbeich
Sponsored by:	DARPA