devel/py-tables: Disable AVX2 CPU optimizations
On amd64 platforms that have a CPU with Haswell New Instructions (AVX2),
related optimizations are automatically compiled in, unless this is
explicitly prevented.

If the package is then used on other amd64 platforms, that have no AVX2
support, this causes problems, because the generated .so files are
incompatible and lead to crashes with SIGILL upon invocation.

Remedy the issue by turning off AVX2 support by default which is also
upstream's default and bump PORTREVISION to force a rebuild.

PR:		258714
Reported by:	dvl (on IRC)
Reviewed by:	wen (maintainer)
Approved by:	portmgr (blanket: build fix)
MFH:		2021Q3