sysutils/calamares: new port, GUI system installer
Calamares is a GUI-based Linux system installer (e.g. for Linux
distro's to get their ISO image installed on a computer). It can
(probably) do FreeBSD as well, although it needs some more work
in order to do the right thing for ZFS (and also the traditional
fetching-base-tarballs bits). Since it is modular and extensible,
adding the FreeBSD-specific bits can be done later. Having a
port seems like a good start (with future ports calamares-extensions
and calamares-freebsd so that a live-ISO-installer could work
with just those packages installed; potentially useful for
helloSystem, FuryBSD and others if they care to collaborate
on GUI-installer-development rather than bespoke bits).