sysutils/kpmcore: new port, KDE partition manager libraries
This is the backend-library for KDE partition-management utilities,
including sysutils/calamares. There is no functional FreeBSD backend,
though, so the port currently builds the dummy backend.

There is rudimentary ZFS support in the libraries, so having
a port available seems like a good start to completing the ZFS
work (at which point it's all-of-a-suddent useful for many
FreeBSD systems).