devel/cmake: Revert "devel/cmake: fix build with python 3.10 as default"
This reverts commit 913219385dfdf72232e6beefe287377cfebdfb04.

Previously build of cmake failed if 3.10 is default version of
Python. And it was worked around by limiting the range of Python to
3.6-3.9. The source of the build error is runtime error of
textproc/py-sphinx with Python 3.10. But it is already fixed and
workaround isn't necessary any more. So revert the commit
corresponding to it.

PR:		258798
Approved by:	adridg (kde@), ygy (mentor)
MFH:		2021Q4
Differential Revision:	https://reviews.freebsd.org/D32282