math/R: Fix build with clang/openmp 13
The mgcv package is one of the recommended packages to be installed with
R.  A problem with mgcv and clang/openmp 13 is that mgcv versions prior
to 1.8-37 place 'include <omp.h>' after 'include <R.h>'.  This causes a
conflict between '#define match' and an openmp pragma line, which uses
match().  This problem was fixed in mgcv 1.8-37, so replace mgcv 1.8-36
with this newer version.

While here, incorporate formatting changes suggested by portclippy and
portfmt.

PR:		258887
Approved by:	dim
Differential Revision:	https://reviews.freebsd.org/D32280