multimedia/qt5-multimedia: avoid gstreamer-gl
While qt5-multimedia **works** with gstreamer-gl, it's unexpected,
and not listed as a dependency. Sabotage the detection of gstreamer-gl
by breaking the include path it looks for.

We might consider adding gstreamer-gl to the dependencies, to
get better overall GL support, but that's a more involved change.

PR:		245909
Reported by:	Martin Birgmeier