net-p2p/py-nicotine-plus: New port: Graphical client for the Soulseek file sharing network
Nicotine+ is a graphical client for the Soulseek peer-to-peer network.
Aims to be a pleasant, free and open source (FOSS) alternative to
the official Soulseek client, providing additional functionality while
keeping current with the Soulseek protocol.
It uses GTK for its graphical user interface, and is written in Python.

WWW: https://github.com/Nicotine-Plus/nicotine-plus/

PR:		246328
Approved by:	arrowd (mentor)
Differential Revision:	https://reviews.freebsd.org/D32290