shells/bash: fix compilation after libncurses split in base
bash actually only requires libtinfo, but up to recent change in
14.0-CURRENT, libtinfo and libncurses were bundled into one single libs,
so linking only to libncurses was not a problem.

By telling bash to link to libtinfo and not to libncurses, it works on
both system where ncurses has been split and those without the split as
there is a libtinfow.so symlink to libncursesw.so on those.