security/linux-bitwarden-cli: Add a new port
The Bitwarden CLI is a powerful, full-featured command-line interface
(CLI) tool to access and manage a Bitwarden vault. The CLI is written
with TypeScript and Node.js and can be run on Windows, macOS, and Linux
distributions. This port is using the Linux binary thanks to the Linux
binary compatibility provided by FreeBSD.

WWW: https://github.com/bitwarden/cli