misc/hicolor-icon-theme: switch to HTTPS site
The original sumbitter adds non-official HTTPS mirror but I'm not sure
where it comes from and whether reliable or not. So I'm doing simple
substitution s/http/https/ for HTTPS available sites.

Most of ports are already switched to to HTTPS site.
misc/hicolor-icon-theme is the only left port.

PR:		239095