devel/desktop-file-utils: use trigger
Instead of polluting all files with a script to be executed post install
convert the handling of mime database into a trigger

This trigger will regenerate the db each time someone install a files
into the concern directory

On removal of the desktop-file-utils port, it will cleanup its cache.

While here convert the post installation script into a lua script.
Note the db is always run at post installation of the port, just to make
sure the cache is up to date, in case the format changes in the futur