Mk/bsd.port.mk: Fix makesum for ports that override MASTER_SITES in make(makesum)
fba040e62bff04323a29d5ad2e21f516b18e9bb4 broke makesum for ports
that conditionally override MASTER_SITES in a make(makesum)
block to simplify maintenance like lang/python-doc-html or
devel/freebsd-sysroot where it is impractical to just statically
list all possible MASTER_SITES. It also broke USES=linux ports
that do the same for DISTFILES.

The breakage with DISTFILES was fixed in
50d2c82e016fd176868cdc6e4befa606fa61c50e by explicitly passing
it to the sub-make fetch. Do the same with MASTER_SITES and
PATCH_SITES so we do not have to workaround this with .MAKEFLAGS
or .export.

PR:		249537
Approved by:	portmgr (implicit)