lang/python3X: Fix library installation
The Python ports install the library libpython3.x.so under $PREFIX/lib,
and they set USE_LDCONFIG, but these libraries are not registered, due
to a missing symlink, and they are not found by `ldconfig -r'.

This commit make them to be registered, and for some reason it helps the
dynamic linker to find them, and this allows to fix an error in
french/aster. It also helps to fix errors in newer releases of math/sage
(not yet ready to be committed due to other problems).

No exp-run, but it has been tested with many ports on several platforms.

PR:		257864
Approved by:	koobs (Python team)
MFH:		2021Q4