Mk/bsd.gcc.mk: Remove USE_GCC=any
We have recommended against USE_GCC=any for a while, and as of more
recently it was completely equivalent to USE_GCC=yes.

With (ancient versions of) GCC hardly available in the base system
of FreeBSD systems at this point, there's unlikely to be a use case
to reintroduce it, so remove the few remaining traces.

PR:		258015